# backend/openapi.yaml
openapi: 3.0.0
info:
  title: CareerAhead
  version: 0.1.0-alpha
  contact:
    name: Pankaj Vishw
    email: vishw.dev.1000@gmail.com
servers:
  - url: http://localhost:{port}/api
    variables:
      port:
        default: "7575"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    signedTokenAuth:
      type: apiKey
      name: token
      in: query
      description: >
        A short-lived signed token used to access specific secure endpoints. 
        The token is generated by the `/api/auth/token/signed-token` endpoint 
        and is valid for 3 minutes. Include it as a query parameter in requests 
        to protected endpoints.
        To get this token you need to have valid access token.
    RazorpaySignature:
      type: apiKey
      in: header
      name: X-Razorpay-Signature
      description: Signature sent by Razorpay to verify the authenticity of the webhook
  schemas:
    UUID:
      description: UUID
      type: string
      format: uuid
      example: be99f5eb-d042-4181-9a59-3708f41a0064
    UserId:
      description: UUID of User
      allOf:
        - $ref: "#/components/schemas/UUID"
    WalletId:
      description: UUID of Wallet
      allOf:
        - $ref: "#/components/schemas/UUID"
    PaymentId:
      description: UUID of Payment
      allOf:
        - $ref: "#/components/schemas/UUID"
    RechargeId:
      description: UUID of Recharge/Order
      allOf:
        - $ref: "#/components/schemas/UUID"
...
    Recharge:
      description: Recharge
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserId"
        wallet:
          $ref: "#/components/schemas/WalletId"
        amount:
          type: integer
          example: 50
        currency:
          type: string
          example: "INR"
        payment_id:
          $ref: "#/components/schemas/PaymentId"
        order_id:
          $ref: "#/components/schemas/RechargeId"
        status:
          type: string
          enum:
            - Created
            - Attempted
            - Paid
            - Failed
...
security:
  - BearerAuth: []
paths:
...
  /wallet/recharges/{recharge_id}/verify:
    post:
      description: Verifies the recharge/order
      operationId: verifyRecharge
      tags:
        - Wallet
      parameters:
        - name: recharge_id
          in: path
          required: true
          schema:
            type: string
            example: order_ifoejfklewfjoei
      security:
        - BearerAuth: []
        - signedTokenAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_id:
                  $ref: "#/components/schemas/PaymentId"
                signature:
                  type: string
                  example: 4DediF-32DFF3090fF#209jFFEDFWEF330903F
      responses:
        "200":
          description: |
            Recharge is verified and money 
            is added to your wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recharge"
...